+++ Z/Data/Array/Checked.hs
@@ -15,7 +15,9 @@ with more informations.
 -}
 module Z.Data.Array.Checked
   ( -- * Arr typeclass re-export
-    A.Arr
+    Arr, MArr
+  , A.singletonArr, A.doubletonArr
+  , modifyIndexArr, insertIndexArr, deleteIndexArr
   , RealWorld
   -- * Boxed array type
   , A.Array(..)
@@ -73,9 +75,12 @@ module Z.Data.Array.Checked
   ) where
 
 import           Control.Exception       (ArrayException (..), throw)
+import           Control.Monad
 import           Control.Monad.Primitive
+import           Control.Monad.ST
 import           Data.Primitive.Types
 import           GHC.Stack
+import           Z.Data.Array          (Arr, MArr)
 import qualified Z.Data.Array          as A
 
 check :: HasCallStack => Bool -> a -> a
@@ -83,18 +88,18 @@ check :: HasCallStack => Bool -> a -> a
 check True  x = x
 check False _ = throw (IndexOutOfBounds $ show callStack)
 
-newArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-       => Int -> m (A.MArr arr s a)
+newArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+       => Int -> m (MArr arr s a)
 newArr n = check  (n>=0) (A.newArr n)
 {-# INLINE newArr #-}
 
-newArrWith :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-           => Int -> a -> m (A.MArr arr s a)
+newArrWith :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+           => Int -> a -> m (MArr arr s a)
 newArrWith n x = check  (n>=0) (A.newArrWith n x)
 {-# INLINE newArrWith #-}
 
-readArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-        => A.MArr arr s a -> Int -> m a
+readArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+        => MArr arr s a -> Int -> m a
 readArr marr i = do
     siz <- A.sizeofMutableArr marr
     check
@@ -102,8 +107,8 @@ readArr marr i = do
         (A.readArr marr i)
 {-# INLINE readArr #-}
 
-writeArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-         => A.MArr arr s a -> Int -> a -> m ()
+writeArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+         => MArr arr s a -> Int -> a -> m ()
 writeArr marr i x = do
     siz <- A.sizeofMutableArr marr
     check
@@ -111,8 +116,8 @@ writeArr marr i x = do
         (A.writeArr marr i x)
 {-# INLINE writeArr #-}
 
-setArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-       => A.MArr arr s a -> Int -> Int -> a -> m ()
+setArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+       => MArr arr s a -> Int -> Int -> a -> m ()
 setArr marr s l x = do
     siz <- A.sizeofMutableArr marr
     check
@@ -120,14 +125,14 @@ setArr marr s l x = do
         (A.setArr marr s l x)
 {-# INLINE setArr #-}
 
-indexArr :: (A.Arr arr a, HasCallStack)
+indexArr :: (Arr arr a, HasCallStack)
          => arr a -> Int -> a
 indexArr arr i = check
     (i>=0 && i<A.sizeofArr arr)
     (A.indexArr arr i)
 {-# INLINE indexArr #-}
 
-indexArr' :: (A.Arr arr a, HasCallStack)
+indexArr' :: (Arr arr a, HasCallStack)
           => arr a -> Int -> (# a #)
 indexArr' arr i =
     if (i>=0 && i<A.sizeofArr arr)
@@ -135,15 +140,15 @@ indexArr' arr i =
     else throw (IndexOutOfBounds $ show callStack)
 {-# INLINE indexArr' #-}
 
-indexArrM :: (A.Arr arr a, Monad m, HasCallStack)
+indexArrM :: (Arr arr a, Monad m, HasCallStack)
           => arr a -> Int -> m a
 indexArrM arr i = check
     (i>=0 && i<A.sizeofArr arr)
     (A.indexArrM arr i)
 {-# INLINE indexArrM #-}
 
-freezeArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-          => A.MArr arr s a -> Int -> Int -> m (arr a)
+freezeArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+          => MArr arr s a -> Int -> Int -> m (arr a)
 freezeArr marr s l = do
     siz <- A.sizeofMutableArr marr
     check
@@ -151,15 +156,15 @@ freezeArr marr s l = do
         (A.freezeArr marr s l)
 {-# INLINE freezeArr #-}
 
-thawArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-        => arr a -> Int -> Int -> m (A.MArr arr s a)
+thawArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+        => arr a -> Int -> Int -> m (MArr arr s a)
 thawArr arr s l = check
     (s>=0 && l>=0 && (s+l)<=A.sizeofArr arr)
     (A.thawArr arr s l)
 {-# INLINE thawArr #-}
 
-copyArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-        => A.MArr arr s a -> Int -> arr a -> Int -> Int -> m ()
+copyArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+        => MArr arr s a -> Int -> arr a -> Int -> Int -> m ()
 copyArr marr s1 arr s2 l = do
     siz <- A.sizeofMutableArr marr
     check
@@ -167,8 +172,8 @@ copyArr marr s1 arr s2 l = do
         (A.copyArr marr s1 arr s2 l)
 {-# INLINE copyArr #-}
 
-copyMutableArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-               => A.MArr arr s a -> Int -> A.MArr arr s a -> Int -> Int -> m ()
+copyMutableArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+               => MArr arr s a -> Int -> MArr arr s a -> Int -> Int -> m ()
 copyMutableArr marr1 s1 marr2 s2 l = do
     siz1 <- A.sizeofMutableArr marr1
     siz2 <- A.sizeofMutableArr marr2
@@ -177,8 +182,8 @@ copyMutableArr marr1 s1 marr2 s2 l = do
         (A.copyMutableArr marr1 s1 marr2 s2 l)
 {-# INLINE copyMutableArr #-}
 
-moveArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-        => A.MArr arr s a -> Int -> A.MArr arr s a -> Int -> Int -> m ()
+moveArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+        => MArr arr s a -> Int -> MArr arr s a -> Int -> Int -> m ()
 moveArr marr1 s1 marr2 s2 l = do
     siz1 <- A.sizeofMutableArr marr1
     siz2 <- A.sizeofMutableArr marr2
@@ -187,15 +192,15 @@ moveArr marr1 s1 marr2 s2 l = do
         (A.copyMutableArr marr1 s1 marr2 s2 l)
 {-# INLINE moveArr #-}
 
-cloneArr :: (A.Arr arr a, HasCallStack)
+cloneArr :: (Arr arr a, HasCallStack)
          => arr a -> Int -> Int -> arr a
 cloneArr arr s l = check
     (s>=0 && l>=0 && (s+l)<=A.sizeofArr arr)
     (A.cloneArr arr s l)
 {-# INLINE cloneArr #-}
 
-cloneMutableArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-                => A.MArr arr s a -> Int -> Int -> m (A.MArr arr s a)
+cloneMutableArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+                => MArr arr s a -> Int -> Int -> m (MArr arr s a)
 cloneMutableArr marr s l = do
     siz <- A.sizeofMutableArr marr
     check
@@ -203,8 +208,8 @@ cloneMutableArr marr s l = do
         (A.cloneMutableArr marr s l)
 {-# INLINE cloneMutableArr #-}
 
-resizeMutableArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-                 => A.MArr arr s a -> Int -> m (A.MArr arr s a)
+resizeMutableArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+                 => MArr arr s a -> Int -> m (MArr arr s a)
 resizeMutableArr marr n = check
     (n>=0)
     (A.resizeMutableArr marr n)
@@ -212,8 +217,8 @@ resizeMutableArr marr n = check
 
 -- | New size should be >= 0, and <= original size.
 --
-shrinkMutableArr :: (A.Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
-                 => A.MArr arr s a -> Int -> m ()
+shrinkMutableArr :: (Arr arr a, PrimMonad m, PrimState m ~ s, HasCallStack)
+                 => MArr arr s a -> Int -> m ()
 shrinkMutableArr marr n = do
     siz <- A.sizeofMutableArr marr
     check
@@ -276,3 +281,48 @@ copyPtrToMutablePrimArray marr s ptr l = do
     check
         (s>=0 && l>=0 && (s+l)<=siz)
         (A.copyPtrToMutablePrimArray marr s ptr l)
+
+--------------------------------------------------------------------------------
+
+modifyIndexArr :: (Arr arr a, HasCallStack) => arr a
+               -> Int    -- ^ offset
+               -> Int    -- ^ length
+               -> Int    -- ^ index in new array
+               -> (a -> a)   -- ^ modify function
+               -> arr a
+{-# INLINE modifyIndexArr #-}
+modifyIndexArr arr off len ix f = runST $ do
+    marr <- A.unsafeThawArr (cloneArr arr off len)
+    !v <- f <$> readArr marr ix
+    writeArr marr ix v
+    A.unsafeFreezeArr marr
+
+-- | Insert an immutable array's element at given index to produce a new array.
+insertIndexArr :: Arr arr a
+               => arr a
+               -> Int        -- ^ offset
+               -> Int        -- ^ length
+               -> Int        -- ^ insert index in new array
+               -> a          -- ^ element to be inserted
+               -> arr a
+{-# INLINE insertIndexArr #-}
+insertIndexArr arr s l i x = runST $ do
+    marr <- newArrWith (l+1) x
+    when (i>0) $ copyArr marr 0 arr s i
+    when (i<l) $ copyArr marr (i+1) arr (i+s) (l-i)
+    A.unsafeFreezeArr marr
+
+-- | Drop an immutable array's element at given index to produce a new array.
+deleteIndexArr :: Arr arr a
+               => arr a
+               -> Int        -- ^ offset
+               -> Int        -- ^ length
+               -> Int        -- ^ drop index in new array
+               -> arr a
+{-# INLINE deleteIndexArr #-}
+deleteIndexArr arr s l i = runST $ do
+    marr <- newArr (l-1)
+    when (i>0) $ copyArr marr 0 arr s i
+    let i' = i+1
+    when (i'<l) $ copyArr marr i arr (i'+s) (l-i')
+    A.unsafeFreezeArr marr