+++ Z/Data/Builder.hs
@@ -34,7 +34,8 @@ module Z.Data.Builder
   , encodePrimLE
   , encodePrimBE
   -- * More builders
-  , stringModifiedUTF8, charModifiedUTF8, stringUTF8, charUTF8, string7, char7, word7, string8, char8, word8, text
+  , stringModifiedUTF8, charModifiedUTF8, stringUTF8
+  , charUTF8, string7, char7, word7, string8, char8, word8, word8N, text
   -- * Numeric builders
   -- ** Integral type formatting
   , IFormat(..)