+++ Z/Data/JSON/Value.hs
@@ -23,7 +23,7 @@ There's no lazy parsers here, every pieces of JSON document will be parsed into 
 
 module Z.Data.JSON.Value
   ( -- * Value type
-    Value(..), get, at
+    Value(..), key, nth
     -- * parse into JSON Value
   , parseValue
   , parseValue'
@@ -48,7 +48,7 @@ import qualified Z.Data.Text.Base       as T
 import           Z.Data.Text.Print     (Print(..))
 import           Z.Data.Vector.Base     as V
 import           Z.Data.Vector.Extra    as V
-import           Z.Data.Vector.FlatMap  as FM
+import           Z.Data.Vector.Search   as V
 import           Z.Foreign
 import           System.IO.Unsafe         (unsafeDupablePerformIO)
 import           Test.QuickCheck.Arbitrary (Arbitrary(..))
@@ -128,22 +128,28 @@ instance Arbitrary Value where
     shrink (Array vs) = V.unpack vs
     shrink _          = []
 
--- | Retrive 'Array' element, return `Null` if 'Value' is not an 'Array' or index not exist.
+-- | Lense for 'Array' element.
 --
-at :: Value -> Int -> Value
-{-# INLINABLE at #-}
-Array vs `at` ix = case vs `indexMaybe` ix of
-    Just v -> v
-    _ -> Null
-_        `at` _  = Null
+-- 1. return `Null` if 'Value' is not an 'Array' or index not exist.
+-- 2. Modify will have no effect if 'Value' is not an 'Array' or index not exist.
+--
+nth :: Functor f => Int -> (Value -> f Value) -> Value -> f Value
+{-# INLINABLE nth #-}
+nth ix f (Array vs) | Just v <- vs `indexMaybe` ix =
+    fmap (\ x -> Array (V.unsafeModifyIndex vs ix (const x))) (f v)
+nth _ f v = fmap (const v) (f Null)
 
--- | Retrive 'Object' element, return `Null` if 'Value' is not an 'Object' or key not exist.
-get :: Value -> T.Text -> Value
-{-# INLINABLE get #-}
-Object m `get` key = case FM.linearSearchR m key of
-    Just v -> v
-    _ -> Null
-_        `get` _   = Null
+-- | Lense for 'Object' element
+--
+-- 1. return `Null` if 'Value' is not an 'Object' or key not exist.
+-- 2. Modify will have no effect if 'Value' is not an 'Object' or key not exist.
+-- 4. On duplicated keys prefer the last one.
+--
+key :: Functor f => T.Text -> (Value -> f Value) -> Value -> f Value
+{-# INLINABLE key #-}
+key k f (Object kvs) | (i, Just (_, v)) <- V.findR ((k ==) . fst) kvs =
+    fmap (\ x -> Object (V.unsafeModifyIndex kvs i (const (k, x)))) (f v)
+key _ f v = fmap (const v) (f Null)
 
 -- | Parse 'Value' without consuming trailing bytes.
 parseValue :: V.Bytes -> (V.Bytes, Either P.ParseError Value)