+++ Z/Data/Vector/FlatMap.hs
@@ -26,8 +26,8 @@ module Z.Data.Vector.FlatMap
   , merge, mergeWithKey'
     -- * fold and traverse
   , foldrWithKey, foldrWithKey', foldlWithKey, foldlWithKey', traverseWithKey
-    -- * search on vectors
-  , linearSearch, linearSearchR, binarySearch
+    -- * binary search on vectors
+  , binarySearch
   ) where
 
 import           Control.DeepSeq
@@ -38,8 +38,9 @@ import qualified Data.Foldable              as Foldable
 import qualified Data.Traversable           as Traversable
 import qualified Data.Semigroup             as Semigroup
 import qualified Data.Monoid                as Monoid
-import qualified Z.Data.Vector.Base as V
-import qualified Z.Data.Vector.Sort as V
+import qualified Z.Data.Vector.Base         as V
+import qualified Z.Data.Vector.Extra        as V
+import qualified Z.Data.Vector.Sort         as V
 import qualified Z.Data.Text.Print          as T
 import           Data.Function              (on)
 import           Data.Bits                  (unsafeShiftR)
@@ -193,45 +194,33 @@ lookup k' (FlatMap (V.Vector arr s l)) = go s (s+l-1)
 -- | /O(N)/ Insert new key value into map, replace old one if key exists.
 insert :: Ord k => k -> v -> FlatMap k v -> FlatMap k v
 {-# INLINE insert #-}
-insert k v (FlatMap vec@(V.Vector arr s l)) =
+insert k v (FlatMap vec) =
     case binarySearch vec k of
-        Left i -> FlatMap (V.create (l+1) (\ marr -> do
-            when (i>0) $ A.copySmallArray marr 0 arr s i
-            A.writeSmallArray marr i (k, v)
-            when (i<l) $ A.copySmallArray marr (i+1) arr (i+s) (l-i)))
-        Right i -> FlatMap (V.Vector (runST (do
-            let arr' = A.cloneSmallArray arr s l
-            marr <- A.unsafeThawSmallArray arr'
-            A.writeSmallArray marr i (k, v)
-            A.unsafeFreezeSmallArray marr)) 0 l)
+        Left i -> FlatMap (V.unsafeInsertIndex vec i (k, v))
+        Right i -> FlatMap (V.unsafeModifyIndex vec i (const (k, v)))
 
 -- | /O(N)/ Delete a key value pair by key.
 delete :: Ord k => k -> FlatMap k v -> FlatMap k v
 {-# INLINE delete #-}
-delete k m@(FlatMap vec@(V.Vector arr s l)) =
+delete k m@(FlatMap vec) =
     case binarySearch vec k of
         Left _ -> m
-        Right i -> FlatMap $ V.create (l-1) (\ marr -> do
-            when (i>0) $ A.copySmallArray marr 0 arr s i
-            let i' = i+1
-            when (i'<l) $ A.copySmallArray marr i arr (i'+s) (l-i'))
+        Right i -> FlatMap (V.unsafeDeleteIndex vec i)
 
 -- | /O(N)/ Modify a value by key.
 --
 -- The value is evaluated to WHNF before writing into map.
 adjust' :: Ord k => (v -> v) -> k -> FlatMap k v -> FlatMap k v
 {-# INLINE adjust' #-}
-adjust' f k m@(FlatMap vec@(V.Vector arr s l)) =
+adjust' f k m@(FlatMap vec) =
     case binarySearch vec k of
         Left _ -> m
-        Right i -> FlatMap $ V.create l (\ marr -> do
-            A.copySmallArray marr 0 arr s l
-            let !v' = f (snd (A.indexSmallArray arr (i+s)))
-            A.writeSmallArray marr i (k, v'))
+        Right i -> FlatMap . V.unsafeModifyIndex vec i $
+            \ (k', v) -> let !v' = f v in (k', v')
 
 -- | /O(n+m)/ Merge two 'FlatMap', prefer right value on key duplication.
 merge :: forall k v. Ord k => FlatMap k v -> FlatMap k v -> FlatMap k v
-{-# INLINE merge #-}
+{-# INLINABLE merge #-}
 merge fmL@(FlatMap (V.Vector arrL sL lL)) fmR@(FlatMap (V.Vector arrR sR lR))
     | null fmL = fmR
     | null fmR = fmL
@@ -356,28 +345,3 @@ binarySearch (V.Vector arr s l) !k' = go s (s+l-1)
             in case k' `compare` k of LT -> go i (mid-1)
                                       GT -> go (mid+1) j
                                       _  -> Right mid
-
---------------------------------------------------------------------------------
-
--- | linear scan search from left to right, return the first one if exist.
-linearSearch :: Ord k => V.Vector (k, v) -> k -> Maybe v
-{-# INLINABLE linearSearch #-}
-linearSearch (V.Vector arr s l) !k' = go s
-  where
-    !end = s + l
-    go !i
-        | i >= end = Nothing
-        | otherwise =
-            let (k, v)  = arr `A.indexSmallArray` i
-            in if k' == k then Just v else go (i+1)
-
--- | linear scan search from right to left, return the first one if exist.
-linearSearchR :: Ord k => V.Vector (k, v) -> k -> Maybe v
-{-# INLINABLE linearSearchR #-}
-linearSearchR (V.Vector arr s l) !k' = go (s+l-1)
-  where
-    go !i
-        | i < s = Nothing
-        | otherwise =
-            let (k, v)  = arr `A.indexSmallArray` i
-            in if k' == k then Just v else go (i-1)